# docs/architecture/openapi-node.yaml
openapi: 3.0.3
info:
  title: DAgents Agent Node API
  version: "0.5"
  description: |
    Agent Node 本地 HTTP API。对外契约以 **Agent** 为中心（1 Agent = 1 主对话）。
    `/v1/sessions*` 已移除（未注册 → 404）；请使用 `/v1/agents/{agent_id}/...`。
servers:
  - url: http://127.0.0.1:18789
    description: 默认本机 Node
paths:
...
  /v1/goals:
    get:
      summary: Node 上的长期任务列表
      responses: {'200': {description: goals, content: {application/json: {schema: {type: object, properties: {goals: {type: array, items: {$ref: '#/components/schemas/Goal'}}}}}}}}
    post:
      summary: 创建长期任务并建立专用会话
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/GoalCreateRequest'}}}
      responses: {'409': {description: configure through the Agent autonomy endpoint}}
  /v1/goals/{goal_id}:
    get:
      parameters: [{name: goal_id, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: goal, content: {application/json: {schema: {$ref: '#/components/schemas/Goal'}}}}}
  /v1/goals/{goal_id}/runs:
    get:
      parameters: [{name: goal_id, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: runs}}
  /v1/goals/{goal_id}/pause:
    post:
      parameters: [{name: goal_id, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: paused}}
  /v1/goals/{goal_id}/resume:
    post:
      parameters: [{name: goal_id, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: resumed}}
  /v1/goals/{goal_id}/stop:
    post:
      parameters: [{name: goal_id, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: stopped}}
  /v1/goals/{goal_id}/wake:
    post:
      parameters: [{name: goal_id, in: path, required: true, schema: {type: string}}]
      responses: {'202': {description: queued, content: {application/json: {schema: {type: object, properties: {run: {$ref: '#/components/schemas/GoalRun'}, delivery_id: {type: string}}}}}}}
...
components:
  schemas:
    GoalCreateRequest:
      type: object
      required: [title, objective, acceptance, agent_id]
      properties:
        title: {type: string, maxLength: 160}
        objective: {type: string}
        acceptance: {type: string}
        agent_id: {type: string}
        max_runs: {type: integer, default: 12, minimum: 1, maximum: 1000}
        token_budget: {type: integer, format: int64, default: 100000}
        turn_token_budget: {type: integer, default: 10000}
        min_wake_interval_seconds: {type: integer, default: 300, minimum: 60}
        expires_at: {type: string, format: date-time}
...
    Goal:
      type: object
      properties:
        id: {type: string}
        title: {type: string}
        objective: {type: string}
        acceptance: {type: string}
        agent_id: {type: string}
        session_id: {type: string, description: 专用会话，不由客户端创建}
        status: {type: string, enum: [active, paused, stopped, completed, waiting, failed]}
        status_reason: {type: string}
        max_runs: {type: integer}
        runs: {type: integer}
        token_budget: {type: integer, format: int64}
        tokens_used: {type: integer, format: int64}
        turn_token_budget: {type: integer}
        min_wake_interval_seconds: {type: integer}
        next_wake_at: {type: string, format: date-time}
        last_checkpoint: {$ref: '#/components/schemas/GoalCheckpoint'}
    GoalCheckpoint:
      type: object
      properties:
        summary: {type: string}
        evidence: {type: array, items: {type: string}}
        artifacts: {type: array, items: {type: string}}
        done: {type: boolean}
        decision:
          type: object
          required: [outcome, next_action, reason, summary]
          properties:
            outcome: {type: string, enum: [progress, no_change, completed, blocked]}
            summary: {type: string}
            next_action: {type: string, enum: [at, event, none, needs_input]}
            reason: {type: string}
            expected_progress: {type: string, description: Required when next_action is at or event.}
            next_wake_at: {type: string, format: date-time}
            next_wake_after_seconds: {type: integer, minimum: 1, maximum: 2678400, description: Mutually exclusive with next_wake_at; resolved by the server clock.}
    GoalRun:
      type: object
      properties:
        id: {type: string}
        goal_id: {type: string}
        session_id: {type: string}
        turn_id: {type: string}
        status: {type: string}
        reason: {type: string}
        tokens_used: {type: integer, format: int64}
...